pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "principle-induction"
version = "0.1.0"
description = "Principle induction on Cellular Automata with GSPO/GRPO on MLX"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "mlx>=0.17.0",
  "mlx-lm>=0.17.0",
  "transformers>=4.40.0",
  "mlx-gen-parity>=0.1.1",
  "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.0.0",
  "ruff>=0.5.0",
  "black>=24.3.0",
]

[project.scripts]
pi-run = "scripts.run_from_config:main"

[tool.setuptools]
packages = ["rl", "tools", "datasets", "scripts"]

